Lazy and wreduce seem correct but cache has dealloc bug

+++ Backend/smart_wrappers/CuCuBLAS/backend_data_wrappers.cu
@@ -30,7 +30,7 @@ void CHLGetMaxCPUmem(size_t* free_mem, size_t* max_mem){
 	return;
 }
 
-long int CHLGetMaxDimSqAsset2D(short Asset2DNum, short dsize, long int step, short loc){
+long int CHLGetMaxDimSqAsset2D(short Asset2DNum, short dsize, long int step, int loc){
 	size_t free_mem, max_mem;
 	if (loc >= 0){
 		int prev_loc; cudaGetDevice(&prev_loc);
@@ -44,7 +44,7 @@ long int CHLGetMaxDimSqAsset2D(short Asset2DNum, short dsize, long int step, sho
 	return maxDim;
 }
 
-long int CHLGetMaxDimAsset1D(short Asset1DNum, short dsize, long int step, short loc){
+long int CHLGetMaxDimAsset1D(short Asset1DNum, short dsize, long int step, int loc){
 	size_t free_mem, max_mem;
 	if (loc >= 0){
 		int prev_loc; cudaGetDevice(&prev_loc);
@@ -106,7 +106,7 @@ void *numa_bind_pin_malloc(long long count, int node_num, short W_flag){
 	return ret;
 }
 
-void* CHLMalloc(long long bytes, short loc, short W_flag){
+void* CHLMalloc(long long bytes, int loc, short W_flag){
   void *ptr = NULL;
   if (loc == CHL_MEMLOCS - 1) {
 #ifdef CLDEBUG
@@ -162,7 +162,7 @@ void numa_pin_free(void *gpuptr, long long bytes) {
 }
 
 
-void CHLFree(void * ptr, long long bytes, short loc){
+void CHLFree(void * ptr, long long bytes, int loc){
 	//if (??? == loc) free(ptr);
 	if ((loc >= CHL_WORKERS && loc < CHL_MEMLOCS) || loc == -1) numa_pin_free(ptr, bytes);
 	else if (loc >= 0 && loc < CHL_WORKERS){
@@ -187,7 +187,7 @@ short CHLGetPtrLoc(void * in_ptr)
 #ifdef CUDA_9_WRAPPER_MESS
   error("CHLGetPtrLoc(9.2 version, ptr =%p): Not supported anymore\n", in_ptr);	
 #else
-	short loc = -42;
+	int loc = -42;
 	cudaPointerAttributes ptr_att;
 	if (cudaSuccess != cudaPointerGetAttributes(&ptr_att, in_ptr)) error("CHLGetPtrLoc(cuda 10+ version, ptr =%p):\
 	//Pointer not visible to CUDA, host alloc or error\n", in_ptr);
@@ -205,7 +205,7 @@ short CHLGetPtrLoc(void * in_ptr)
 
 short CHLGetPtrAdvLoc(void * in_ptr, long long dim1, long long dim2, int elemSize)
 {
-	short loc = -42;
+	int loc = -42;
 	cudaPointerAttributes ptr_att;
 	if (cudaSuccess != cudaPointerGetAttributes(&ptr_att, in_ptr)) error("CHLGetPtrAdvLoc(cuda 10+ version, ptr =%p):"
 	"Pointer not visible to CUDA, host alloc or error\n", in_ptr);
@@ -231,7 +231,7 @@ short CHLGetPtrAdvLoc(void * in_ptr, long long dim1, long long dim2, int elemSiz
 	else return CHLGetPtrLoc(in_ptr);
 }
 
-void CHLMemcpy(void* dest, void* src, long long bytes, short loc_dest, short loc_src)
+void CHLMemcpy(void* dest, void* src, long long bytes, int loc_dest, int loc_src)
 {
 	massert(loc_dest >= -1 && loc_dest < CHL_MEMLOCS, "CHLMemcpy: Invalid destination device: %d\n", loc_dest);
 	massert(loc_src >= -1 && loc_src < CHL_MEMLOCS, "CHLMemcpy: Invalid source device: %d\n", loc_src);
@@ -251,7 +251,7 @@ void CHLMemcpy(void* dest, void* src, long long bytes, short loc_dest, short loc
 	CHLSyncCheckErr();
 }
 
-/*void CHLMemcpyAsync(void* dest, void* src, long long bytes, short loc_dest, short loc_src, CQueue_p transfer_queue)
+/*void CHLMemcpyAsync(void* dest, void* src, long long bytes, int loc_dest, int loc_src, CQueue_p transfer_queue)
 {
 
 	cudaStream_t stream = *((cudaStream_t*) transfer_queue->backend_queue_ptr);
@@ -273,7 +273,7 @@ void CHLMemcpy(void* dest, void* src, long long bytes, short loc_dest, short loc
 	//CHLSyncCheckErr();
 }*/
 
-void CHLMemcpy2D(void* dest, long int ldest, void* src, long int ldsrc, long int rows, long int cols, short elemSize, short loc_dest, short loc_src){
+void CHLMemcpy2D(void* dest, long int ldest, void* src, long int ldsrc, long int rows, long int cols, short elemSize, int loc_dest, int loc_src){
 	short lvl = 6;
 #ifdef DDEBUG
 	lprintf(lvl, "CHLMemcpy2D(dest=%p, ldest =%zu, src=%p, ldsrc = %zu, rows = %zu, cols = %zu, elemsize = %d, loc_dest = %d, loc_src = %d)\n",
@@ -301,7 +301,7 @@ void CHLMemcpy2D(void* dest, long int ldest, void* src, long int ldsrc, long int
 }
 
 /*
-void CoCMempy2DAsyncWrap3D(void* dest, long int ldest, void* src, long int ldsrc, long int rows, long int cols, short elemSize, short loc_dest, short loc_src, CQueue_p transfer_queue){
+void CoCMempy2DAsyncWrap3D(void* dest, long int ldest, void* src, long int ldsrc, long int rows, long int cols, short elemSize, int loc_dest, int loc_src, CQueue_p transfer_queue){
 	// Convert 2d input (as CHLMemcpy2DAsync) to 3D for ...reasons.
 	enum cudaMemcpyKind kind = cudaMemcpyDefault;
 	cudaStream_t stream = *((cudaStream_t*) transfer_queue->backend_queue_ptr);
@@ -313,7 +313,7 @@ void CoCMempy2DAsyncWrap3D(void* dest, long int ldest, void* src, long int ldsrc
 }*/
 
 template<typename VALUETYPE>
-void CHLVecInit(VALUETYPE *vec, long long length, int seed, short loc)
+void CHLVecInit(VALUETYPE *vec, long long length, int seed, int loc)
 {
   if (!vec) error("CHLVecInit: vec is not allocated (correctly)\n");
   if (loc < -1  || loc >= CHL_MEMLOCS) error("CHLVecInit: Invalid device id/location %d\n", loc);
@@ -345,5 +345,5 @@ void CHLVecInit(VALUETYPE *vec, long long length, int seed, short loc)
   CHLSyncCheckErr();
 }
 
-template void CHLVecInit<double>(double *vec, long long length, int seed, short loc);
-template void CHLVecInit<float>(float *vec, long long length, int seed, short loc);
+template void CHLVecInit<double>(double *vec, long long length, int seed, int loc);
+template void CHLVecInit<float>(float *vec, long long length, int seed, int loc);
